# openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ISS Visibility API",
    "version": "1.7.0-lr7",
    "description": "ЛР7: API для авторизации, услуг, заявок и m-m точек заявки."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ServiceListPageResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServiceResponse"
            }
          },
          "total": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "page_size": {
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ]
      },
      "ServiceResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "elevation": {
            "type": "integer"
          },
          "timezone": {
            "type": "string"
          },
          "best_time": {
            "type": "string"
          },
          "light_pollution": {
            "type": "string"
          },
          "weather_conditions": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "image_url": {
            "type": "string"
          },
          "video_url": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "country",
          "latitude",
          "longitude",
          "elevation",
          "timezone"
        ]
      },
      "ISSPositionListItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "status_ru": {
            "type": "string"
          },
          "iss_position_determination_results_count": {
            "type": "integer",
            "description": "Количество результатов: 0 пока заявка не завершена, при completed — число точек"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "formed_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "observation_date": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "total_visibility": {
            "type": "string"
          },
          "creator_login": {
            "type": "string"
          },
          "moderator_login": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "status_ru",
          "iss_position_determination_results_count",
          "created_at",
          "total_visibility",
          "creator_login"
        ]
      },
...
    }
  },
  "paths": {
...
    "/api/services": {
      "get": {
        "summary": "Список услуг (пагинация)",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "country",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "timezone",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "min_elevation",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "max_elevation",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "viewed",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Недавно просмотренные — ответ массивом (без пагинации)"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "exclude_id",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Пагинированный список услуг (или массив при viewed=1)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceListPageResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/iss-positions": {
      "get": {
        "summary": "Список заявок",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "formed_from",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "2026-05-05"
            }
          },
          {
            "name": "formed_to",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "2026-05-05"
            }
          },
          {
            "name": "observation_points_count",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            },
            "description": "Фильтр по числу точек наблюдения в заявке"
          }
        ],
        "responses": {
          "200": {
            "description": "Список заявок",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ISSPositionListItem"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}